+++ EPMS.Interfaces/Repository/IEmployeeRepository.cs
@@ -9,8 +9,9 @@ using EPMS.Models.ResponseModels;
 
 namespace EPMS.Interfaces.Repository
 {
-    public interface IEmployeeRepository : IBaseRepository<Employee, int>
+    public interface IEmployeeRepository : IBaseRepository<Employee, long>
     {
         EmployeeResponse GetAllEmployees(EmployeeSearchRequset employeeSearchRequset);
+
     }
 }
